feat: min_width code block property
## Details

Adds `code -> min_width` which can be used to set a minimim width for
block width code blocks. This functions in addition to right padding.
Makes it easy to create uniform width code block sections.

Other changes are a change to the handler add methods to flatten the type
structure of marks and shorten things a bit.