fix: better support for same buffer in multiple windows
# Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/122

When using zen mode plugin 2 instances of the same buffer are open in
different windows. However this plugin only picks up the first window,
thus interactions with the second window did nothing.

To fix this rather than using the `bufwinid` method to go from buffer
to window we now instead get the current window and pass that as input
to the update method along with the buffer. That way the buffer can be
associated with any number of windows but events will still apply to
whatever the active window is.