feat: add border option to headings
# Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/123

This add a `heading -> border` field that is a boolean. This is slightly
different to code borders since it does not seem likely that different
styles will be implemented, more just off and on. The characters added
above and below are also configurable as new options.

The logic attempts to overlay the border on empty lines but will use
virtual text lines if the line is not available. Must also be aware of
the previous heading that was overlayed since that effectively makes a
line non empty as well.

This is essentially a port of the `fat_headlines` option in
headlines.nvim: https://github.com/lukas-reineke/headlines.nvim