feat: heading width based on level
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/126

The `heading -> width` option is now allowed to be `block`, `width`, or
an array of these 2 values. If it is an array we index into it using a
clamp based on the heading level.

Requires a new kind of validation logic but not too bad, seems to cover
all cases. Also required moving list class to generics, which was easy.
Move all union strings to type alias annotations so exact values can be
defined once and reused as needed.