fix: account for folds when computing visible range
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/138

After the change to only parse text visible in the view port we
essentially broke for users who use folds. This is because the range
calculation only used the size of the window and the start.

This change iterates through the rows and determines if they are hidden
by a fold. Otherwise the logic is mostly the same except using lines
that are not folded instead of just the lines.