chore: add row class for unit testing
## Details

Add a Row class used in unit tests which tracks the current row and
implements a get and increment method.

This allows new lines to be inserted in any test files without needing
to modify the row numbers manually for all marks below the added one.

This is because all marks are already sorted by row so adding an
additional increment call will automatically shift all lower marks down.
This only works if the same row instance is used to generate all the
expected marks which is why it has been added to all test classes as the
only way to get the row value, no hard coded integers.