feat: independent code block language padding
## Details

Original PR: https://github.com/MeanderingProgrammer/render-markdown.nvim/pull/131

The idea of this this change is that there is a slight difference in
behavior for left aligned vs right aligned language for code blocks.

- right: language is put at edge of block, essentially not padded
- left: language is indented along with rest of block, essentially padded

To make the behavior more consistent the language will no longer be
padded for left aligned, unlike before, so it will not be effected by
left_pad like the code block body.

A new field has been added `code.language_pad` which will add padding
specific to the language section. To have the same behavior as before
for left aligned this value should be set to the same value as for
`code.left_pad`.

The padding is applied relative to this position value.

- left: language is left padded, moving it to the right
- right: language is right padded, moving it to the left

This allows more customization for users while allowing for the same
behavior as before if that is preferred. Though not technically fully
backwards compatible as getting the same behavior as before requires
setting an additional option value.

Co-authored-by: Biggybi <Biggybi@users.noreply.github.com>