fix: indented tables and tables with no spaces in cells
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/142

This resolves 2 separate issues.

The first is for tables that are indented some amount space over. In
such cases the top & bottom full sections as well as the delimitter
would be misaligned since they did not take into account any leading
spaces and generate full lines rather than changing individual parts
which is the case for rows. To fix this add logic to get the amount of
leading space and shift these components over accordingly.

The second is for tables where the contents of a cell went all the way
to the right pipe with no spaces. This would cause the padding to be
inserted in the middle of the content rather than at the end. To fix
this adjust the column associated with padding to no longer subtract one
off. Use a priority value of 0 for the padding marks so the pipe marks
pickup the padding as well, otherwise they would end up ignoring it.