feat: make custom link order deterministic
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/146

Currently we pick the first matching custom link when selecting an icon
for a destination. However because our table is keyed on strings we
iterate through it using pairs and the order of pairs is not
deterministic.

To fix this collect all matching patterns into a list. Then pick the
match with the longest string pattern length. The idea being that longer
patterns are more specific and likely are what the user prefers to use.
This allows doing fallbacks, like www.example.com for a specific website
then www. for all other websites.