feat: setext heading
## Details

Adds support for the alternative markdown setext headings.
Uses the same settings from the existing heading configuration.

Some refactoring needed to happen to do this effectively since these
headings can be any number of lines so per row for loops were added
in a few places.

Supports all the same features and behaves in the same way as atx
headings with 2 exceptions:

- borders are not supported: this seems redundant to do since setext
  headings always have a bottom border effectively
- bottom text is concealed: the bottom line of these headings always has
  either equals signs or dashes, to make things look prettier this line
  gets concealed, can be replaced in the future with something else if
  users ask