fix: only create foreground when inversing highlight group
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/154

When creating an inverse highlight group for borders we currently swap
both the foreground and background. However there's never really a need
to do this since all we care about is applying the original background
as the new foreground.

Be removing the background property we should avoid issues where users
set a foreground as well but expect the result to blend in with the
overall background.

This also more closely mirrors the goal of the logic.