feat: improve health check for obsidian.nvim conflict
## Details

Currently when using obsidian.nvim we provide blanket advice to disable
the UI and allow the warning to be supressed with a configuration option.

This is ok, but ideally we would actually do the check to make sure the
UI is disabled and avoid the error entirely if it is.

This change does exactly this by pulling the UI enable option from the
obsidian.nvim client. This is very fragile and has no guarantee of
working long term, but it should do well enough for now. It's likely as
fragile as the advice provided before.

This does remove the `acknowledge_conflicts` option which in some ways
is not backwards compatible. However this option only plays a role in
the health check warning and assuming the advice was followed there
should continue to be no errors to users in the health check. As such I
will say this is a non breaking change, but anyone using this option
should remove it. Keeping it causes no problems, except in the health
check for the config.