perf: store callouts and checkboxes in table keyed on raw value
## Details

Currently resolving the callout or checkbox for a particular
shortcut_link requires iterating through the configured values one at a
time. This is slower than it needs to be since we can pre-compute all
the normalized raw values once, store that as the key to the table, then
do a single lookup into the table when needed, skipping the for loop.

To do this add an additional buffer config type that gets returned from
state. This new config will store the normalized component mappings
along with all of the other values already present. This will only
run once per buffer.