feat!: allow skipping heading indents so body is offset
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/161

There is a breaking change:

- `indent.skip` -> `indent.skip_level`

I added this a few hours ago so hopefully not used by many people but
noting it down in either case.

The feature request was to add an indent to section bodies offsetting
them from the section titles.

This is done, but in the other way. Instead indenting headings can be
skipped using `indent.skip_heading`. This makes it easier to calculate
reasonable offsets based on parent sections. It also avoids inconsistent
definitions for level based indentation. For instance if you're working
in an H2 then add an H4 but you still want the heading to line up with
the content of the H2 it gets strange. You might be able to count
nesting levels but this puts us back in a state we moved away from.

By skipping headings instead all the sematics and definitions stay the
same. And by simply not indenting the heading it will naturally line up
with the content of the parent section.