feat: various logging improvements
## Details

Related issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/165

It is currently difficult to figure out why a user is experiencing
problems with this plugin just not working. To improve this we need to
make some pretty large changes to how the logger works, since it
currently it only runs at a per render level. Meaning most information
is lost, i.e. we overwrite the file and we don't log anything until
we're attempting to render a file, which will not help figure out
something not working in general.

To imrove this:

- log information when attempting to attach to a buffer along with
  reasons for any skipped buffers
- doing the previous items requires updating when items are flushed to
  disk since there is no longer a convenient start and end, so instead:
  - When 1000 items have been added without being written
  - When neovim is closed
  - If the `log` command is triggered by the user
- logger is now additive, meaning we keep appending logs to the same file
- logic to prevent file from getting too large, currently hard coded to
  5 MB, after which the file is cleared
- adds `log` subcommand to quickly open log file in a new tab, similar to
  `MasonLog` and `LspLog`
- input to log is now a `vararg` of messages rather than a single item

Unrelated changes:

- renames `logger` module to `log`
- Moves `RenderMarkdown` command creation logic out of plugin directory
  and into its own module
- Small change to `in_section` logic for nodes