feat: allow render_modes to be a boolean
## Details

Allow the render modes parameter to be a boolean rather than a list of
modes. When it is a boolean the mode of the current buffer stops playing
a role and rendering either always occurs or never occurs. Using false
does not make sense in this case but by setting it to true one can avoid
adding all the modes to the list.

Made the main buffer Config a class so that logic to determine things
like should render and handling the different types can be centralized.