feat: disable rendering in diff mode
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/169

Was cut as a question however there was no way to accomplish this before.

The behavior is made global with no way to disable it for now. I can add
one if requested however I don't see that happening as if you are
looking at a diff you probably want to see the underlying text and not a
rendered approximation.

Change is rather small, only adding a check on the `diff` window option
when deciding what the next state of the buffer should be.

We also need to listen to the `DiffUpdated` event so that rendering is
correctly disabled when switching between a diff mode view of the buffer
and a standard view.