feat: reload runtime highlights on color scheme change
## Details

Add a ColorScheme autocommand which regenerates cached colors created at
runtime.

To do this the cache now needs to store all the information needed to
re-create the colors rather than just the colors that have already been
created.

Moves the VimLeave autocommand out of manager and directly into the
setup method of the log module. This keeps the manager module coupled to
buffer level events and allows the log module to manage its lifecycle in
a self contained way.