chore(refactor): share overlap logic between debug & render
## Details

Moves top level render method in Extmark out into ui module which
decides based on overlap whether to show or hide each mark.

Moves namespace and buf out of Extmark, instead passing it in to methods
as needed.

Updates debug method to use the same overlap method as the main update
call that does rendering. Effectively means that everything hidden on
the current line is what gets returned when debugging + those marks
which are not concealable and would be hidden based on range overlap.