fix: update window options on alternate buffer switch
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/177

Reported as an issue when using telescope but is a general problem when
opening a different file in the current window then switching back using
the alternate buffer `<C-6>` keymap.

This occurs because changing the buffer in the current window resets the
local window options. However we have logic to only set window options
on a state change. Essentially we were not capturing the idea that
options are reset when opening a different buffer in the current window.

To fix this remove the logic that gates setting window options on a
state change and instead always update the window options based on the
current state. This shouldn't have much performance impact to users.

Minor other changes:

- Add distribtion input to issue templates
- Add `info` log level which provides some more high level information
  than `error` but not as much noise as `debug`
- Remove level specific methods from `log` and instead take level as an
  input argument and compare levels as integers to decide whether to log