fix: account for columns before text when getting window width
## Details

The current window width gets the entire available space for a window,
but not the usable space. This was okay for getting a fill width but is
not great when computing offsets to center blocks.

The fix is to use the `getwininfo` `textoff` property which conveniently
tells us the width of foldcolumn, signcolumn and numbers before text,
exactly what we're looking for.

To avoid re-computing this cache the value after it is first accessed so
we can re-use it within the same parse cycle.