feat: heading padding options allow arrays, opt in always virtual border
## Details

Request 1: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/182

This change adds support for 2 features.

The first is around the padding options for headings. For each of:

- `left_margin`
- `left_pad`
- `right_pad`
- `min_width`

We now accept an array of numbers in addition to a single number. When
the value is an array we index it using the level and we clamp the last
value. This is the same behavior we support for `width`.

Request 2: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/183

This request is simply to allow using virtual lines for heading borders
all the time rather than trying to overlay borders on empty lines.

This can now be opted in to by setting `border_virtual` to `true`.