fix: not updating when window scrolled
\#\# Details

The rendering is not updated when the window is scrolled while the
cursor is fixed (can be reproduces using `<C-y>`, `<C-e>`, `zz`, `zt`,
`zb` to scroll the window). This leaves new elements that becomes
visible in current window undecorated.

To fix this simply add 'WinScrolled' event to the updating events.