chore(refactor): use list instead of array naming, custom iter for types, indent latex
## Details

Mostly a refactor but does include a latex functional change.

Anywhere we use array replace it with list since that's what various
neovim functions use.

Create an `iter` module with `map` and `filter` setup in such that the
lua language server automatically infers function arguments and out
correctly.

For latex pad the output such that it lines up with the column of the
input formula, should be mildly nicer than always at the start.