Skip debounce when <= 0
## Details

Main functional change is to directly skip debounce when it is set to 0.
This should be the behavior when using timers anyway but this makes it a
more obvious behavior.

Also adds a check for vim.uv and falls back to vim.loop when checking
file size.

Minor other change, add a `log_runtime` option to allow us to print
runtime of update function. Easier to see performance changes when
developing if they just get printed. Users are welcome to it as well if
they are curious.