fix: only set option value if it changes
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/186

In order to fix https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/177
the window option setting logic was updated to always modify window
options rather than on a state transition. The reason was to handle out
of band state transitions for a buffer that would cause the state of the
window options to be changed without the inner states knowledge. As a
result the window options were not being properly updated since
effectively no state transition occured.

While this does seem to have almost no performance penalty it does seem
to reset some kind of state in neovim that causes some kinds of motions
to be have differently, such as moving between ends of lines as reported
in the issue.

To fix this rather than always updating the option value, get the
current value and only set the new value if it is different.

In this way we properly handle out of band transitions since we're
relying on the actual current option value and not some internal state.
And since we don't unnecessarily change the value we don't mess with any
internal neovim state and allow motions to behave as they normally would.