fix: handle offset conceal nodes
## Details

Currently when computing the range of a conceal highlight we always use
the entire range of the node.

However if an offset directive is specified then the range in the
metadata should be used instead.

This change accomplishes this by adding a series of fallbacks when
computing conceal information. First by checking the metadata range,
then the metadata[node_id] range, and finally falling back to the node
range of neither of these exist.