fix: indented table border
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/191

When calulating the leading spaces for a table row to determine how much
to indent a border only the node text was used. However if tables are
indented enough under a list item a leading continuation node eats some
of the spaces. This does not happen consistently between the heading and
other rows resulting in the number of spaces appearing different. When
this occurs we assume the table is not properly aligned on the left and
do not render a border.

To fix this caclulate the number of leadings spaces using the entire
line for the row and not just the node text.

Other unrelated changes:

- Move treesitter queries out of treesitter module and into where they
  are actually used
- Use the treesitter module to cache parsed queries to avoid parsing
  more than once per unique query
- This also avoid needing to do a vim.schedule in our state setup,
  instead the queries end up being naturally lazily evaluated