fix: use virtual text for heading border at start & end
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/187

When using smoothscrolling and textwrap on long lines pasting text
would cause a visual bug where text was duplicated.

This ended up being because when a heading is the last line in a file we
attempt to use an overlay to put the border below it.

I assumed I had written the logic to use virtual text on both ends of a
document however it turns out virtual text was only used at the top most
line as a special case and not at the bottom.

The fix for this bug is then straightforward and we now validate a line
exists and is empty rather than not existing meaning empty.

While this specific issue is fixed it does seem there are generally some
issues with smoothscrolling, textwrap, and virtual lines.