feat: allow callouts to override quote icon
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/194

This adds an optional `quote_icon` field for each callout.

If provided this value will be used instead of the `quote.icon` value.

This means callouts can specify a custom value for the right bar of the quote.
Functions very similar to callout highlight except it is always optional.

Most changes to make this happen are related to splitting the type for
custom checkboxes and custom callouts, which before used the same class.

Now there is a different class for callouts and checkboxes which
impacted context and config normalization.