feat: margin for paragraphs
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/200

Adds a new paragraph configuration which has 2 properties in addition to
the standard `enabled` field.

- `min_width`: allows you to specify a minimum width for each paragraph
  to improve general alignment when using long and short paragraphs
- `left_margin`: this functions in the same way as heading and code
  block margin where a value < 1 is treated as a fraction of the
  available window space and values >= 1 are just then number of cells
  to use for margin.

We use margin rather than padding since paragraphs do not support a
background so padding would not have any visual distenction.

Includes a minor refactor to allow `list.clamp` to resolve either a
single value or an array of values rather than only having this availble
in the heading module. That's still the only place that it's used but in
general I think it makes sense to have it more available.