feat: support on attach callback to integrate external latex support
## Details

Adds `on.attach` configurable function callback. This can be used to
trigger events or other functionality that helps render a markdown
document.

The concrete use case has to do with triggering an external latex
renderer since it does not generate any autocommands and requires the
user to call it via function, keymap, or otherwise.

While there is currently only a single callback point on attach this can
potentially be expanded in the future to trigger other callbacks at
other points in the rendering lifecycle. I do not have any concrete use
cases for this now so will leave this up to feature requests.