feat: user configurable anti conceal per component
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/204

Previously the decoration elements that are hidden on the cursor line
was not something the user had direct control over.

The default behavior was to never hide virtual lines, spacing, signs or
code backgrounds. All other decorations on the line would always be
hidden if the user kept anti conceal enabled.

For the time being virtual lines and spacing remain permanently enabled.

However signs, code backgrounds and most other elements that by default
would be hidden are entirely within the users control.

This is all managed through the new `anti_conceal.ignore` property. This
table is a mapping from element name to a boolean. A value of `true` for
an element means it will not be concealed when the cursor enters the
line. A value of `false` behaves the same as no value being set and will
be hidden. A table is used rather than a list so that when users modify
values the defaults remain the same as they were unless explicitely changed.

Signs and code backgrounds set their values to `true` to remain ignored by
default as before, however now users have the ability to change this
behavior if they would like to hide signs and backgrounds.

Additionally most other decorations that remain concealable by default
like before can now be ignored if the user sets their value to true. A
complete list of possible values is provided in the default configuration
in the README.