fix: window offset bottom calculation
## Details

When rendering a window we use the visible range + some buffer above and
below. This lets us avoid work when scrolling in an already computed
range and can allow us to avoid being debounced when rendering is fairly snappy.

However the math for getting the bottom was not right and we essentially
only added this offset above and not below.

Fix is fairly easy, updated unit tests accordingly.