feat: align table cells according to indicators
## Details

This adds full support for rendering table cells using alignment
indicators and shifting text as needed to accomplish this. This works
with both the `padded` and `trimmed` cell styles.

An additional property `pipe_table.padding` was also added which
defaults to 1 and is used to determine the minimum padding needed within
each table cell. This works even if the cell underneath has no padding as
well as being shifted any amount in any direction. Centering in narrow
cells can sometimes look strange as the inability to use half widths
can make the off by one seem larger, but there's really no way around this
and most of the time it looks great. I'm pretty happy with this one :)

This also modifies the rendering of the center indicator in the
delimiter row to be 2 dashes on both ends of the cell rather than one in
the middle, I think this looks a little better.